DaliLM-Backend-main/DaliLM-Backend-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aioodbc==0.5.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
azure-core==1.34.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
databricks-sql-connector==4.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
fonttools==4.58.1
frozenlist==1.6.2
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-experimental==0.3.4
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.44
lz4==4.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.2.2
openai==1.84.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymssql==2.3.4
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
thrift==0.20.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
zstandard==0.23.0